chore: refactor useThrottle
