+++ packages/ui/src/composables/index.ts
@@ -33,7 +33,6 @@ export * from './useTimer'
 export * from './useTrackBy'
 export * from './useTranslation'
 export * from './useTrapFocus'
-export * from './useThrottle'
 export * from './useValidation'
 export * from './std/ssr/useWindow'
 export * from './std/browser/useWindowSize'
@@ -65,3 +64,5 @@ export * from './std/internal/useSelectorTemplateRef'
 export * from './std/browser/useElementFocusedKeyboard'
 export * from './std/event/useLongPress'
 export * from './std/utils/useDebounceFn'
+export * from './std/utils/useThrottleFn'
+export * from './std/utils/useThrottleValue'
