+++ packages/ui/src/composables/std/index.ts
@@ -23,3 +23,5 @@ export * from './internal/useElementTemplateRef'
 export * from './internal/useSelectorTemplateRef'
 export * from './browser/useElementFocusedKeyboard'
 export * from './utils/useDebounceFn'
+export * from './utils/useThrottleFn'
+export * from './utils/useThrottleValue'
